# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel Gemini",
        "description": "API for interacting with Gemini AI services",
        "contact": {
            "email": "dushbehera05@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "API Server"
        }
    ],
    "paths": {
        "/api/gemini/generate": {
            "post": {
                "tags": [
                    "Gemini"
                ],
                "summary": "Generate a response using Gemini",
                "operationId": "831dfab21f51744f0e6496ba3ee4f154",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "message"
                                ],
                                "properties": {
                                    "message": {
                                        "type": "string",
                                        "example": "Hello, how are you?"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "response": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Message is required"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/gemini/generate-structured": {
            "post": {
                "tags": [
                    "Gemini"
                ],
                "summary": "Generate a structured response using Gemini",
                "operationId": "1ed750ffcc0e78a80bac85909b54ec86",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "message",
                                    "schema_type"
                                ],
                                "properties": {
                                    "message": {
                                        "type": "string",
                                        "example": "Explain quantum physics"
                                    },
                                    "schema_type": {
                                        "type": "string",
                                        "example": "educational"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "response": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/gemini/generate-with-image": {
            "post": {
                "tags": [
                    "Gemini"
                ],
                "summary": "Generate response with image analysis",
                "operationId": "0b04453a55e2a9d36527ea3ea57a6152",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "message",
                                    "image"
                                ],
                                "properties": {
                                    "message": {
                                        "type": "string",
                                        "example": "What's in this image?"
                                    },
                                    "image": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "response": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/gemini/analyze-with-file": {
            "post": {
                "tags": [
                    "Gemini"
                ],
                "summary": "Analyze uploaded file (PDF/MP4)",
                "operationId": "defe02cc0c3ab2cbf2334990a143b525",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "message",
                                    "file"
                                ],
                                "properties": {
                                    "message": {
                                        "type": "string",
                                        "example": "Analyze this document"
                                    },
                                    "file": {
                                        "description": "PDF or MP4 file",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful analysis",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "response": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/gemini/stream": {
            "post": {
                "tags": [
                    "Gemini"
                ],
                "summary": "Stream content from Gemini in real-time",
                "description": "Returns server-sent events stream. Best tested with cURL or EventSource, not typical API clients. Example: curl -N -X POST http://127.0.0.1:8000/api/gemini/stream -H 'Content-Type: application/json' -d '{\"message\": \"Write a story on Gemini.\"}' The -N flag disables buffering for real-time streaming.",
                "operationId": "8a93d598eee0f8b4c2ec605c06d83e7a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "message"
                                ],
                                "properties": {
                                    "message": {
                                        "type": "string",
                                        "example": "Tell me a story"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Server-sent events stream (use cURL or EventSource)",
                        "content": {
                            "text/event-stream": {}
                        }
                    }
                }
            }
        },
...
    },
...
}